+++ Job-Hunt/app/Http/Controllers/Admin/PackageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Package;
+use App\Models\Order;
 
 class PackageController extends Controller
 {
@@ -80,6 +81,12 @@ class PackageController extends Controller
 
     public function delete($id)
     {
+          $check = Order::where('package_id',$id)->count();
+          if($check>0) 
+          {
+            return redirect()->back()->with('error', 'You can not delete this item, because this is used in another place.');
+          }
+          
           $package  = Package::find($id);
           $package ->delete();
